add documents, rework tests and delete unnecessary classes.
