+++ flink-python/pyflink/table/tests/test_table_environment_api.py
@@ -15,6 +15,7 @@
 # #  See the License for the specific language governing permissions and
 # # limitations under the License.
 ################################################################################
+import datetime
 import os
 import tempfile
 
@@ -23,7 +24,179 @@ from pyflink.table.table_environment import TableEnvironment
 from pyflink.table.table_config import TableConfig
 from pyflink.table.table_sink import CsvTableSink
 from pyflink.table.types import DataTypes
-from pyflink.testing.test_case_utils import PyFlinkBatchTableTestCase
+from pyflink.testing import source_sink_utils
+from pyflink.testing.test_case_utils import PyFlinkBatchTableTestCase, PyFlinkStreamTableTestCase
+
+
+class StreamTableEnvironmentTests(PyFlinkStreamTableTestCase):
+
+    def test_register_scan(self):
+        source_path = os.path.join(self.tempdir + '/streaming.csv')
+        field_names = ["a", "b", "c"]
+        field_types = [DataTypes.INT, DataTypes.STRING, DataTypes.STRING]
+        data = [(1, "Hi", "Hello"), (2, "Hello", "Hello")]
+        csv_source = self.prepare_csv_source(source_path, data, field_types, field_names)
+
+        t_env = self.t_env
+        t_env.register_table_source("Source", csv_source)
+        result = t_env.scan("Source")
+
+        field_names = ["a", "b", "c"]
+        field_types = [DataTypes.INT, DataTypes.STRING, DataTypes.STRING]
+        t_env.register_table_sink(
+            "Results",
+            field_names, field_types, source_sink_utils.TestAppendSink())
+        result.insert_into("Results")
+        t_env.execute()
+        actual = source_sink_utils.results()
+        expected = ['1,Hi,Hello', '2,Hello,Hello']
+        self.assert_equals(actual, expected)
+
+    def test_register_table_source_sink(self):
+        source_path = os.path.join(self.tempdir + '/streaming.csv')
+        field_names = ["a", "b", "c"]
+        field_types = [DataTypes.INT, DataTypes.STRING, DataTypes.STRING]
+        data = [(1, "Hi", "Hello")]
+        csv_source = self.prepare_csv_source(source_path, data, field_types, field_names)
+
+        t_env = self.t_env
+        t_env.register_table_source("Orders", csv_source)
+
+        t_env.register_table_sink(
+            "Sinks",
+            field_names, field_types, source_sink_utils.TestAppendSink())
+
+        t_env.scan("Orders").insert_into("Sinks")
+        t_env.execute()
+
+        actual = source_sink_utils.results()
+        expected = ['1,Hi,Hello']
+        self.assert_equals(actual, expected)
+
+    def test_from_table_source(self):
+        source_path = os.path.join(self.tempdir + '/streaming.csv')
+        field_names = ["a", "b", "c"]
+        field_types = [DataTypes.INT, DataTypes.STRING, DataTypes.STRING]
+        data = [(1, "Hi", "Hello"), (2, "Hi", "Hello")]
+        csv_source = self.prepare_csv_source(source_path, data, field_types, field_names)
+
+        t_env = self.t_env
+        source = t_env.from_table_source(csv_source)
+
+        t_env.register_table_sink(
+            "Sinks",
+            field_names, field_types, source_sink_utils.TestAppendSink())
+
+        source.insert_into("Sinks")
+        t_env.execute()
+
+        actual = source_sink_utils.results()
+        expected = ['1,Hi,Hello', '2,Hi,Hello']
+        self.assert_equals(actual, expected)
+
+    def test_list_tables(self):
+        source_path = os.path.join(self.tempdir + '/streaming.csv')
+        field_names = ["a", "b", "c"]
+        field_types = [DataTypes.INT, DataTypes.STRING, DataTypes.STRING]
+        data = []
+        csv_source = self.prepare_csv_source(source_path, data, field_types, field_names)
+        t_env = self.t_env
+        t_env.register_table_source("Orders", csv_source)
+
+        t_env.register_table_sink(
+            "Sinks",
+            field_names, field_types, source_sink_utils.TestAppendSink())
+
+        t_env.register_table_sink(
+            "Results",
+            field_names, field_types, source_sink_utils.TestAppendSink())
+
+        actual = t_env.list_tables()
+        expected = ['Orders', 'Results', 'Sinks']
+        self.assert_equals(actual, expected)
+
+    def test_explain(self):
+        source_path = os.path.join(self.tempdir + '/streaming.csv')
+        field_names = ["a", "b", "c"]
+        field_types = [DataTypes.INT, DataTypes.STRING, DataTypes.STRING]
+        data = []
+        csv_source = self.prepare_csv_source(source_path, data, field_types, field_names)
+
+        t_env = self.t_env
+        t_env.register_table_source("Source", csv_source)
+        source = t_env.scan("Source")
+        result = source.alias("a, b, c").select("1 + a, b, c")
+
+        actual = t_env.explain(result)
+        print(actual)
+        assert isinstance(actual, str) or isinstance(actual, unicode)
+
+    def test_sql_query_update(self):
+        source_path = os.path.join(self.tempdir + '/streaming.csv')
+        field_names = ["a", "b", "c"]
+        field_types = [DataTypes.INT, DataTypes.STRING, DataTypes.STRING]
+        data = [(1, "Hi", "Hello"), (2, "Hello", "Hello")]
+        csv_source = self.prepare_csv_source(source_path, data, field_types, field_names)
+
+        t_env = self.t_env
+        t_env.register_table_source("Source", csv_source)
+        source = t_env.scan("Source")
+        result = t_env.sql_query("select a + 1, b, c from %s" % source)
+        t_env.register_table("results", result)
+
+        t_env.register_table_sink(
+            "sinks",
+            field_names, field_types, source_sink_utils.TestAppendSink())
+
+        query_config = t_env.query_config()
+        query_config.with_idle_state_retention_time(
+            datetime.timedelta(days=1), datetime.timedelta(days=2))
+        assert query_config.get_max_idle_state_retention_time() == 2 * 24 * 3600 * 1000
+        assert query_config.get_min_idle_state_retention_time() == 24 * 3600 * 1000
+
+        t_env.sql_update("insert into sinks select * from results", query_config)
+        t_env.execute("test_sql_job")
+
+        actual = source_sink_utils.results()
+        expected = ['2,Hi,Hello', '3,Hello,Hello']
+        self.assert_equals(actual, expected)
+
+        t_env.sql_update("insert into sinks select * from results")
+        t_env.execute()
+
+        actual = source_sink_utils.results()
+        expected = ['2,Hi,Hello', '3,Hello,Hello']
+        self.assert_equals(actual, expected)
+
+    def test_table_config(self):
+        t_env = self.t_env
+        table_config = t_env.get_config()
+        table_config.parallelism = 4
+        table_config.null_check = True
+        table_config.max_generated_code_length = 64000
+        table_config.timezone = "Asia/Shanghai"
+        table_config.is_stream = True
+        assert table_config.parallelism == 4
+        assert table_config.null_check is True
+        assert table_config.max_generated_code_length == 64000
+        assert table_config.timezone == "Asia/Shanghai"
+        assert table_config.is_stream is True
+
+    def test_create_table_environment(self):
+        table_config = TableConfig.Builder()\
+            .set_parallelism(2)\
+            .set_max_generated_code_length(32000)\
+            .set_null_check(False)\
+            .set_timezone("Asia/Shanghai")\
+            .as_streaming_execution()\
+            .build()
+        t_env = TableEnvironment.create(table_config)
+        readed_table_config = t_env.get_config()
+        assert readed_table_config.parallelism == 2
+        assert readed_table_config.null_check is False
+        assert readed_table_config.max_generated_code_length == 32000
+        assert readed_table_config.timezone == "Asia/Shanghai"
+        assert readed_table_config.is_stream is True
 
 
 class BatchTableEnvironmentTests(PyFlinkBatchTableTestCase):
